fix-aula03-finish

+++ server/src/routes/memories.ts
@@ -4,8 +4,17 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  //preHandler significa antes de executar a funcao de cada uma das rotas eu quero que verifique se o usuario esta autenticado
+  app.addHook('preHandler', async(request) =>{
+    //a funcao verifica na requisicao que o front ta fazendo para essa rota, ta vindo o token (jwt) dele, se nao estiver, ele nao deixa prosseguir
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy:{
         createdAt: 'asc',
       }
@@ -20,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
 
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -33,6 +42,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       }
     })
+
+    if (!memory.isPublic && memory.userId != request.user.sub){
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -51,14 +64,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '0f9744b2-3097-4499-a84a-646da71421a1',
+        userId: request.user.sub,
       }
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -73,7 +86,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-   const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where:{
+        id,
+      }
+    })
+
+    if (memory.userId != request.user.sub){
+        return reply.status(401).send()
+    }
+
+   memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -87,12 +110,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const{ id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where:{
+        id,
+      }
+    })
+
+    if (memory.userId != request.user.sub){
+        return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
